+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/ExplicitUsageApi.kt
@@ -1,9 +1,9 @@
 package com.windea.breezeframework.core.annotations.api
 
-/**稳定的api。*/
+/**需要显式使用的api。*/
 @MustBeDocumented
 @Retention(AnnotationRetention.BINARY)
-annotation class StableApi(
+annotation class ExplicitUsageApi(
 	/**备注信息。*/
-	val value: String = ""
+	val message: String = ""
 )