+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/ReferenceApi.kt
@@ -1,15 +1,15 @@
-package com.windea.breezeframework.core.annotations.marks
+package com.windea.breezeframework.core.annotations.api
 
 import org.intellij.lang.annotations.*
 import kotlin.annotation.AnnotationTarget.*
 
-/**标注这个实现参考自哪里。*/
+/**带有外部引用的Api。*/
 @MustBeDocumented
 @Repeatable
 @Retention(AnnotationRetention.SOURCE)
 @Target(CLASS, PROPERTY, FUNCTION, FILE)
-annotation class Reference(
-	/**引用地址。*/
+annotation class ReferenceApi(
+	/**备注信息。*/
 	@Language("Markdown")
-	val value: String
+	val message: String
 )