+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/enums/core/LetterCase.kt
@@ -115,14 +115,14 @@ enum class LetterCase(
 		{ it.splitToWordList('-') },
 		{ it.joinToString("-") }
 	),
-	@ExplicitApi
+	@ExplicitUsageApi
 	`camelCase_AllowUnderscore`(
 		//allow: lower word first, capitalized/upper words remain , $ surrounds a word, numbers, _ may repeat & at begin
 		"""_*\$?[a-z]+(?:_*(?:\$?[A-Z][a-z]+\$?|\$?[A-Z]+\$?|\d+))+""".toRegex(),
 		{ it.toWords().splitToWordList() },
 		{ it.joinToString("") { S -> S.firstCharToUpperCaseOnly() }.firstCharToLowerCase() }
 	),
-	@ExplicitApi
+	@ExplicitUsageApi
 	`PascalCase_AllowUnderscore`(
 		//allow: lower word first, capitalized/upper words remain , $ surrounds a word, numbers, _ may repeat & at begin
 		"""_*\$?(?:[A-Z][a-z]+|[A-Z]+)(?:_*(?:\$?[A-Z][a-z]+\$?|\$?[A-Z]+\$?|\d+))+""".toRegex(),