+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/FileExtensions.kt
@@ -1,7 +1,6 @@
 package com.windea.breezeframework.core.extensions
 
 import com.windea.breezeframework.core.annotations.api.*
-import com.windea.breezeframework.core.annotations.marks.*
 import java.io.*
 import java.net.*
 
@@ -11,12 +10,12 @@ val File.shotName: String get() = this.nameWithoutExtension
 /**得到文件的MIME类型。*/
 val File.mimeType: String? get() = URLConnection.guessContentTypeFromName(this.name)
 
+//https://zhidao.baidu.com/question/2078337860385272108.html
+//https://www.oschina.net/question/112255_44552
+//https://www.iteye.com/blog/chinacheng-841270
+//https://www.cnblogs.com/chenglc/p/7117847.html
 /**得到文件的真实MIME类型。*/
 @TrickImplementationApi("Performance and accuracy problem.")
-@Reference("<https://zhidao.baidu.com/question/2078337860385272108.html>")
-@Reference("<https://www.oschina.net/question/112255_44552>")
-@Reference("<https://www.iteye.com/blog/chinacheng-841270>")
-@Reference("<https://www.cnblogs.com/chenglc/p/7117847.html>")
 val File.actualMimeType: String?
 	get() = URLConnection.guessContentTypeFromStream(this.inputStream())